Breaking Change - ISE Hostname -> ISE URL
This change requires that the object is instantiated using the full ISE
URL (https://hostname:port) instead of just the hostname as previous.

This change allows the caller to specify different URLs for different
environments.
In our particular use case, we required to hit a Mock ISE server running
http on port 8000.